+++ app/controllers/concerns/authorization.rb
@@ -9,7 +9,7 @@ module Authorization
     private
 
     def user_not_authorized
-      flash[:danger] = "You do not have permission to access this."
+      flash[:danger] = 'You do not have permission to access this.'
       redirect_to(request.referer || root_path)
     end
   end