+++ app/helpers/admin/users_helper.rb
@@ -1,7 +1,7 @@
 module Admin
   module UsersHelper
     def user_roles
-      User.roles.keys.map { |role| [ role.titleize, role ] }
+      User.roles.keys.map { |role| [role.titleize, role] }
     end
   end
 end