+++ app/jobs/monthly_report_job.rb
@@ -4,10 +4,12 @@ class MonthlyReportJob < ApplicationJob
   def perform
     User.find_each do |user|
       if user.report_frequency == 'monthly'
-        MonthlyReportMailer.with(user: user, eaten_products: user.eaten_products
-        .where(eaten_at: (Date.today - 1).all_month)).monthly_report.deliver_now
+        eaten_products = user.eaten_products.where(eaten_at: Date.yesterday.all_month)
+        MonthlyReportMailer.with(
+          user: user,
+          eaten_products: eaten_products
+        ).monthly_report.deliver_now
       end
     end
   end
 end
-