+++ app/models/user.rb
@@ -13,7 +13,7 @@ class User < ApplicationRecord
   before_validation :remove_redundant_chars!
 
   validates :email, presence: true, uniqueness: true,
-  format: { with: /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z\d\-]+)*\.[a-z]+\z/i }
+                    format: { with: /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z\d\-]+)*\.[a-z]+\z/i }
 
   def guest?
     false