+++ app/services/user_bulk_service.rb
@@ -22,8 +22,7 @@ class UserBulkService < ApplicationService
       User.new(name: cells[0],
                email: cells[1],
                password: cells[2],
-               password_confirmation: cells[2]
-      )
+               password_confirmation: cells[2])
     end
   end
 end