+++ test/models/eaten_product_meal_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class EatenProductMealTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end